Improve nickname mention accuracy
I sent a message that started with "jed21611: ", because that was his
current nickname, and frontdesk sent me a PM saying that "jed2161" is
not currently in the room, and he'll be messaged when he is.

This change should fix that bug, only recognizing a nickname in a message if
it ends with a colon or a space.

I'd like to write some tests for this function but I'm so unfamiliar with go
so far and I haven't looked how to do that yet.